examples/servers/simple-notebook/pyproject.toml
[project]
name = "mcp-simple-notebook"
version = "0.1.0"
description = "A simple MCP server demonstrating scope-based OAuth authentication with notebook tools"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Anthropic, PBC." }]
license = { text = "MIT" }
dependencies = [
    "anyio>=4.5",
    "click>=8.2.0",
    "httpx>=0.27",
    "mcp",
    "pydantic>=2.0",
    "pydantic-settings>=2.5.2",
    "sse-starlette>=1.6.1",
    "uvicorn>=0.23.1; sys_platform != 'emscripten'",
]

[project.scripts]
mcp-simple-notebook-rs = "mcp_simple_notebook.server:main"
mcp-simple-notebook-as = "mcp_simple_notebook.auth_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_simple_notebook"]

[dependency-groups]
dev = ["pyright>=1.1.391", "pytest>=8.4.4", "ruff>=0.14.11"]
